+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/GeneralCrudService.java
@@ -119,41 +119,41 @@ public class GeneralCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     @Override
-    public K insert(E element) throws ServiceException {
+    public K insert(E entity) throws ServiceException {
         try {
-            return internalInsert(element);
+            return internalInsert(entity);
         } catch (Exception e) {
             throw ServiceExceptionHelper.logParse("插入实体时发生异常", exceptionLogLevel, e, sem);
         }
     }
 
-    private K internalInsert(E element) throws Exception {
-        if (Objects.isNull(element.getKey())) {
-            element.setKey(keyGenerator.generate());
-        } else if (internalExists(element.getKey())) {
+    private K internalInsert(E entity) throws Exception {
+        if (Objects.isNull(entity.getKey())) {
+            entity.setKey(keyGenerator.generate());
+        } else if (internalExists(entity.getKey())) {
             throw new ServiceException(ServiceExceptionCodes.ENTITY_EXISTED);
         }
-        K key = dao.insert(element);
-        cache.push(element, cacheTimeout);
+        K key = dao.insert(entity);
+        cache.push(entity, cacheTimeout);
         return key;
     }
 
     @Override
-    public void update(E element) throws ServiceException {
+    public void update(E entity) throws ServiceException {
         try {
-            internalUpdate(element);
+            internalUpdate(entity);
         } catch (Exception e) {
             throw ServiceExceptionHelper.logParse("更新实体时发生异常", exceptionLogLevel, e, sem);
         }
     }
 
-    private void internalUpdate(E element) throws Exception {
-        if (!internalExists(element.getKey())) {
+    private void internalUpdate(E entity) throws Exception {
+        if (!internalExists(entity.getKey())) {
             throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
         }
 
-        dao.update(element);
-        cache.push(element, cacheTimeout);
+        dao.update(entity);
+        cache.push(entity, cacheTimeout);
     }
 
     @Override
@@ -186,10 +186,10 @@ public class GeneralCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     @Override
-    public K insertIfNotExists(E element) throws ServiceException {
+    public K insertIfNotExists(E entity) throws ServiceException {
         try {
-            if (Objects.isNull(element.getKey()) || !internalExists(element.getKey())) {
-                return internalInsert(element);
+            if (Objects.isNull(entity.getKey()) || !internalExists(entity.getKey())) {
+                return internalInsert(entity);
             }
             return null;
         } catch (Exception e) {
@@ -198,10 +198,10 @@ public class GeneralCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     @Override
-    public void updateIfExists(E element) throws ServiceException {
+    public void updateIfExists(E entity) throws ServiceException {
         try {
-            if (internalExists(element.getKey())) {
-                internalUpdate(element);
+            if (internalExists(entity.getKey())) {
+                internalUpdate(entity);
             }
         } catch (Exception e) {
             throw ServiceExceptionHelper.logParse("更新实体时发生异常", exceptionLogLevel, e, sem);
@@ -220,12 +220,12 @@ public class GeneralCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     @Override
-    public K insertOrUpdate(E element) throws ServiceException {
+    public K insertOrUpdate(E entity) throws ServiceException {
         try {
-            if (Objects.isNull(element.getKey()) || !internalExists(element.getKey())) {
-                return internalInsert(element);
+            if (Objects.isNull(entity.getKey()) || !internalExists(entity.getKey())) {
+                return internalInsert(entity);
             } else {
-                internalUpdate(element);
+                internalUpdate(entity);
                 return null;
             }
         } catch (Exception e) {