+++ subgrade-stack/src/main/java/com/dwarfeng/subgrade/stack/service/BatchCrudService.java
@@ -15,28 +15,34 @@ import java.util.List;
 public interface BatchCrudService<K extends Key, E extends Entity<K>> extends CrudService<K, E> {
 
     /**
-     * 查询指定的主键是否全部存在。
+     * 批量查询实体是否存在。
      *
-     * @param keys 指定的主键组成的列表。
+     * <p>
+     * 如果指定的实体键列表中的所有键对应的实体全部存在，则返回 <code>true</code>；否则返回 <code>false</code>。
+     *
+     * @param keys 实体的键组成的列表。
      * @return 指定的主键是否全部存在。
      * @throws ServiceException 服务异常。
      */
     boolean allExists(List<K> keys) throws ServiceException;
 
     /**
-     * 查询指定的主键是否全部不存在。
+     * 批量查询实体是否存在。
      *
-     * @param keys 指定的主键组成的列表。
+     * <p>
+     * 如果指定的实体键列表中的所有键对应的实体全部不存在，则返回 <code>true</code>；否则返回 <code>false</code>。
+     *
+     * @param keys 实体的键组成的列表。
      * @return 指定的主键是否全部不存在。
      * @throws ServiceException 服务异常。
      */
     boolean nonExists(List<K> keys) throws ServiceException;
 
     /**
-     * 批量获取指定的键对应的元素。
+     * 批量获取实体。
      *
-     * @param keys 指定的键组成的列表。
-     * @return 指定的键对应的元素组成的列表。
+     * @param keys 实体的键组成的列表。
+     * @return 实体的键对应的实体组成的列表。
      * @throws ServiceException 服务异常。
      */
     List<E> batchGet(List<K> keys) throws ServiceException;
@@ -44,42 +50,72 @@ public interface BatchCrudService<K extends Key, E extends Entity<K>> extends Cr
     /**
      * 批量插入实体。
      *
-     * @param elements 实体组成的列表。
+     * <p>
+     * 对于批量操作的每一个实体，当实体的主键不存在时，会尝试生成一个新的主键。
+     *
+     * <p>
+     * 对于批量操作的每一个实体，需要注意的是，插入行为可能会改变实体的主键，即使实体的主键已经被设置。
+     *
+     * <p>
+     * 批量操作的每一个实体，插入完成后，对于实体的主键会被设置为插入后的主键。
+     *
+     * @param entities 实体组成的列表
      * @return 插入的实体对应的主键。
      * @throws ServiceException 服务异常。
      */
-    List<K> batchInsert(List<E> elements) throws ServiceException;
+    List<K> batchInsert(List<E> entities) throws ServiceException;
 
     /**
      * 批量更新实体。
      *
-     * @param elements 实体组成的列表。
+     * @param entities 实体组成的列表
      * @throws ServiceException 服务异常。
      */
-    void batchUpdate(List<E> elements) throws ServiceException;
+    void batchUpdate(List<E> entities) throws ServiceException;
 
     /**
      * 批量删除实体。
      *
-     * @param keys 实体组成的列表。
+     * @param keys 实体的键组成的列表。
      * @throws ServiceException 服务异常。
      */
     void batchDelete(List<K> keys) throws ServiceException;
 
     /**
-     * 批量获取指定且存在的键对应的元素。
+     * 批量获取实体。
      *
-     * @param keys 指定的键组成的列表。
-     * @return 指定且存在的键对应的元素组成的列表，元素的个数小于等于键列表的个数。
+     * <p>
+     * 对于批量操作中的每一个键，如果对应的实体存在，则获取该实体；否则不做任何处理。<br>
+     * 将所有获取的实体组成一个列表返回。
+     *
+     * <p>
+     * 返回的列表中的实体的个数小于等于键列表的个数。
+     *
+     * @param keys 实体的键组成的列表。
+     * @return 指定且存在的键对应的实体组成的列表，实体的个数小于等于键列表的个数。
      * @throws ServiceException 服务异常。
      */
     List<E> batchGetIfExists(List<K> keys) throws ServiceException;
 
     /**
-     * 如果不存在指定的元素，则插入。
+     * 批量插入实体。
      *
-     * @param elements 指定的元素组成的集合。
-     * @return 所有被插入元素的主键，小于或等于 elements 的个数。
+     * <p>
+     * 对于批量操作的每一个实体，如果实体不存在，则插入，并获取插入后的主键；否则不做任何处理。<br>
+     * 将所有插入的实体的主键组成一个列表返回。
+     *
+     * <p>
+     * 对于批量操作的每一个实体，当实体不存在，需要插入实体时，需要注意以下事项：
+     * <ul>
+     *   <li>插入行为可能会改变实体的主键，即使实体的主键已经被设置。</li>
+     *   <li>插入完成后，实体的主键会被设置为插入后的主键。</li>
+     * </ul>
+     *
+     * <p>
+     * 返回的列表中的主键的个数小于等于实体列表的个数。
+     *
+     * @param entities 实体组成的列表
+     * @return 所有被插入实体的主键，小于或等于 entities 的个数。
      * @throws ServiceException 服务异常。
      * @see #batchInsertIfNotExists(List)
      * @deprecated 该方法不符合语义，不建议使用。请使用 {@link #batchInsertIfNotExists(List)}。
@@ -90,41 +126,75 @@ public interface BatchCrudService<K extends Key, E extends Entity<K>> extends Cr
      */
     @SuppressWarnings("DeprecatedIsStillUsed")
     @Deprecated
-    List<K> batchInsertIfExists(List<E> elements) throws ServiceException;
+    List<K> batchInsertIfExists(List<E> entities) throws ServiceException;
 
     /**
-     * 如果不存在指定的元素，则插入。
+     * 批量插入实体。
      *
-     * @param elements 指定的元素组成的集合。
-     * @return 所有被插入元素的主键，小于或等于 elements 的个数。
+     * <p>
+     * 对于批量操作的每一个实体，如果实体不存在，则插入，并获取插入后的主键；否则不做任何处理。<br>
+     * 将所有插入的实体的主键组成一个列表返回。
+     *
+     * <p>
+     * 对于批量操作的每一个实体，当实体不存在，需要插入实体时，需要注意以下事项：
+     * <ul>
+     *   <li>插入行为可能会改变实体的主键，即使实体的主键已经被设置。</li>
+     *   <li>插入完成后，实体的主键会被设置为插入后的主键。</li>
+     * </ul>
+     *
+     * <p>
+     * 返回的列表中的主键的个数小于等于实体列表的个数。
+     *
+     * @param entities 实体组成的列表
+     * @return 所有被插入实体的主键，小于或等于 entities 的个数。
      * @throws ServiceException 服务异常。
      */
-    default List<K> batchInsertIfNotExists(List<E> elements) throws ServiceException {
-        return batchInsertIfExists(elements);
+    default List<K> batchInsertIfNotExists(List<E> entities) throws ServiceException {
+        return batchInsertIfExists(entities);
     }
 
     /**
-     * 如果存在指定的元素，则更新。
+     * 批量更新实体。
      *
-     * @param elements 指定的元素组成的集合。
+     * <p>
+     * 对于批量操作的每一个实体，如果实体存在，则更新；否则不做任何处理。
+     *
+     * @param entities 实体组成的列表
      * @throws ServiceException 服务异常。
      */
-    void batchUpdateIfExists(List<E> elements) throws ServiceException;
+    void batchUpdateIfExists(List<E> entities) throws ServiceException;
 
     /**
-     * 如果存在指定的元素，则删除。
+     * 批量删除实体。
      *
-     * @param keys 指定的元素对应的键组成的集合。
+     * <p>
+     * 对于批量操作的每一个键，如果对应的实体存在，则删除；否则不做任何处理。
+     *
+     * @param keys 实体的键组成的列表。
      * @throws ServiceException 服务异常。
      */
     void batchDeleteIfExists(List<K> keys) throws ServiceException;
 
     /**
-     * 如果元素不存在，则插入；存在，则更新。
+     * 批量插入或更新实体。
      *
-     * @param elements 指定的元素组成的集合。
-     * @return 所有被插入元素的主键，小于或等于 elements 的个数。
+     * <p>
+     * 对于批量操作的每一个实体，如果实体不存在，则插入实体，并获取插入后的主键；否则更新实体。<br>
+     * 将所有插入或更新的实体的主键组成一个列表返回。
+     *
+     * <p>
+     * 对于批量操作的每一个实体，当实体不存在，需要插入实体时，需要注意以下事项：
+     * <ul>
+     *   <li>插入行为可能会改变实体的主键，即使实体的主键已经被设置。</li>
+     *   <li>插入完成后，实体的主键会被设置为插入后的主键。</li>
+     * </ul>
+     *
+     * <p>
+     * 返回的列表中的主键的个数小于等于实体列表的个数。
+     *
+     * @param entities 实体组成的列表
+     * @return 所有被插入实体的主键，小于或等于 entities 的个数。
      * @throws ServiceException 服务异常。
      */
-    List<K> batchInsertOrUpdate(List<E> elements) throws ServiceException;
+    List<K> batchInsertOrUpdate(List<E> entities) throws ServiceException;
 }