+++ subgrade-stack/src/main/java/com/dwarfeng/subgrade/stack/service/WriteService.java
@@ -21,16 +21,31 @@ import com.dwarfeng.subgrade.stack.exception.ServiceException;
 public interface WriteService<E extends Entity<?>> extends Service {
 
     /**
-     * 写入指定的元素。
-     * <p>
-     * 写入的元素应该遵守如下的规则：
-     * <pre>
-     * 1. 元素的主键允许为 null。
-     * 2. 主键不为 null 的元素必须确保之前不存在。
-     * </pre>
+     * 写入实体。
      *
-     * @param element 指定的元素。
+     * <p>
+     * 为保证写入的效率，该方法默认待写入的实体是合法的，因此不会对实体进行合法性检查。
+     *
+     * <p>
+     * 待写入的实体需要满足如下的约定：
+     * <ul>
+     *   <li>实体主键允许为 <code>null</code>。</li>
+     *   <li>如果实体的主键不为 <code>null</code>，则实体必须确保之前不存在。</li>
+     * </ul>
+     *
+     * <p>
+     * 如果传入的实体不合法，该方法可能会抛出异常，其具体行为由实现决定。
+     *
+     * <p>
+     * 需要注意的是，该方法写入数据后，不会返回数据的主键，即使数据的主键是 <code>null</code>，或者是在写入时被改变。<br>
+     * 因此，在项目开发中，需要使用其它的服务，确保写入的实体是可以被查询到的，例如可以使用：
+     * <ul>
+     *   <li>{@link EntireLookupService}</li>
+     *   <li>{@link PresetLookupService}</li>
+     * </ul>
+     *
+     * @param entity 实体。
      * @throws ServiceException 服务异常。
      */
-    void write(E element) throws ServiceException;
+    void write(E entity) throws ServiceException;
 }